server/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
openai
python-dotenv
